Relocate request-language classification cluster (batch 11)
Phase 2 batch 11 for Issue #682. Moves the verifier-adjacent request-
language classification cluster (~140 LOC) out of turn.rs into
verifier_orchestration.rs. Removes all 4 remaining
super::turn:: back-imports from verifier_orchestration.rs.

Moved (turn.rs -> verifier_orchestration.rs):

  Pure helpers:
    - decision_target_path (#[cfg(test)], 12 LOC) — projects the
      workspace-target half of a VerifierRepairDecision so the
      MissingVerifierJob safeguard can re-validate scope when no
      real RepairJob is attached. Previously back-imported from
      turn.rs.
    - missing_verifier_setup_hint_for_request (25 LOC) — returns
      language-specific config hint for the no-verifier note.
      Previously back-imported from turn.rs by both
      task_contract_no_verifier_note and verifier_setup_policy_message.
    - synthesized_missing_implementation_target_path_for_request
      (34 LOC) — request-derived default implementation target path
    - synthesized_missing_test_target_path_for_request (23 LOC) —
      request-derived default test target path
    - test_target_path_compatible_with_request (12 LOC) — gate that
      checks a hint path's extension against the language family
      implied by the request

  Pattern-match helpers:
    - lower_contains_any (3 LOC)
    - request_contains_any (3 LOC)
    - request_matches_family (8 LOC)

  Language-pattern consts (10 entries, ~11 LOC):
    - PYTHON_REQUEST_PATTERNS / PYTHON_REQUEST_JA_PATTERNS /
      PYTHON_TEST_REQUEST_PATTERNS
    - RUST_REQUEST_PATTERNS / RUST_REQUEST_JA_PATTERNS /
      RUST_LIBRARY_REQUEST_PATTERNS / RUST_LIBRARY_REQUEST_JA_PATTERNS
    - TYPESCRIPT_REQUEST_PATTERNS / JAVASCRIPT_REQUEST_PATTERNS /
      NODE_REQUEST_PATTERNS

All 8 fns + 10 consts promoted to pub(super). decision_target_path
retains its #[cfg(test)] attribute. The cluster is now fully self-
contained inside verifier_orchestration.rs.

Adjustments:

  * verifier_orchestration.rs: removed missing_verifier_setup_hint_for_request +
    decision_target_path from the super::turn:: back-imports (no
    longer needed — both are now defined locally). The
    super::turn:: block is now only the 2 TASK_CONTRACT_VERIFIER_*
    consts.

  * turn.rs: extended use super::verifier_orchestration::{...} with
    synthesized_missing_implementation_target_path_for_request /
    synthesized_missing_test_target_path_for_request /
    test_target_path_compatible_with_request (the 3 production
    callers from `impl Agent` methods at L10849/10887/10892).
    Rewrote test-mod `super::missing_verifier_setup_hint_for_request(`
    + `super::decision_target_path(` references to
    super::super::verifier_orchestration::*.

Metrics: turn.rs 33,149 -> 33,004 (-145);
verifier_orchestration.rs 1,920 -> 2,063 (+143). Cumulative Issue
#682 progress: turn.rs 34,958 -> 33,004 (-1,954, ~65% of Phase 2
target). Acceptance gap: 33,004 -> 32,000 = ~1,004 LOC remaining.
